simplify file structure
