--- pyroll/core/profile/hookimpls/__init__.py
@@ -1,2 +1,0 @@
-from . import equivalent_rectangle
-from . import shape
