--- pyroll/core/transport/hookimpls/__init__.py
@@ -1,1 +1,0 @@
-from . import total_softening
